@@ -4,8 +4,7 @@ import 'package:flutter/material.dart';
 class Convertbutton extends StatefulWidget {
   final HomeController homeController;
 
-  const Convertbutton({Key? key, required this.homeController})
-      : super(key: key);
+  const Convertbutton({super.key, required this.homeController});
 
   @override
   State<Convertbutton> createState() => _ConvertbuttonState();
@@ -16,9 +15,14 @@ class _ConvertbuttonState extends State<Convertbutton> {
   Widget build(BuildContext context) {
     return ElevatedButton(
       style: ButtonStyle(
-        backgroundColor: MaterialStateProperty.all(
-          const Color.fromARGB(108, 136, 115, 59),
+        backgroundColor: WidgetStateProperty.all(
+          const Color.fromRGBO(255, 255, 255, 1),
         ),
+        overlayColor: const WidgetStatePropertyAll(
+          Color.fromARGB(255, 102, 100, 100),
+        ),
+        foregroundColor:
+            const WidgetStatePropertyAll(Color.fromARGB(255, 255, 255, 255)),
       ),
       onPressed: () {
         setState(() {
@@ -28,7 +32,7 @@ class _ConvertbuttonState extends State<Convertbutton> {
       child: const Text(
         "CONVERTER",
         style: TextStyle(
-          color: Colors.amber,
+          color: Color.fromARGB(255, 0, 0, 0),
         ),
       ),
     );
