@@ -1,4 +1,5 @@
 import 'package:conversor_de_moedas/App/HomePageView/Controllers/home_controller.dart';
+import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:conversor_de_moedas/App/HomePageView/Elements/ButtonsAndTextF/dropdown_buttons.dart';
 import 'package:flutter/material.dart';
 
@@ -12,6 +13,8 @@ class MoedasCont extends StatefulWidget {
 }
 
 class MoedasContState extends State<MoedasCont> {
+  final List<String> _currencyCodes = ["BRL", "USD", "EUR", "GBP", "BTC"];
+
   @override
   Widget build(BuildContext context) {
     return Container(
@@ -19,16 +22,18 @@ class MoedasContState extends State<MoedasCont> {
       child: Column(
         children: [
           SizedBox(
-            width: 300,
-            height: 40,
+            width: 160.w,
+            height: 45.h,
             child: Row(
               children: [
                 Expanded(
                   flex: 1,
                   child: DropDownB(
-                    onSelected: (value) {
+                    onSelected: (String? newValue) {
                       setState(() {
-                        widget.homeController.alteraMoedasFrom(value);
+                        widget.homeController.alteraMoedasFrom(
+                            _currencyCodes.indexOf(
+                                newValue ?? "BRL")); 
                       });
                     },
                     positionDropdown: widget.homeController.indexMoedasFrom,
@@ -37,8 +42,10 @@ class MoedasContState extends State<MoedasCont> {
                 Expanded(
                   flex: 2,
                   child: Container(
-                    height: 300,
-                    margin: const EdgeInsets.only(left: 20, bottom: 8),
+                    height: 28.h,
+                    margin: const EdgeInsets.only(
+                      left: 20,
+                    ),
                     child: TextField(
                       controller: widget.homeController.fromTextMoedas,
                       style: const TextStyle(
@@ -46,13 +53,19 @@ class MoedasContState extends State<MoedasCont> {
                           color: Colors.amber),
                       textAlign: TextAlign.center,
                       decoration: const InputDecoration(
+                        fillColor: Colors.black,
+                        focusedBorder: UnderlineInputBorder(
+                          borderSide: BorderSide(
+                            color: Color.fromARGB(255, 194, 194, 192),
+                          ),
+                        ),
                         enabledBorder: UnderlineInputBorder(
                           borderSide: BorderSide(
-                            color: Colors.amber,
+                            color: Color.fromARGB(255, 194, 194, 192),
                           ),
                         ),
                       ),
-                      cursorColor: Colors.amber,
+                      cursorColor: const Color.fromARGB(255, 194, 194, 192),
                     ),
                   ),
                 ),
